fix(schema,mcp-server,cli,acp-server): declare detailProblem on the wire; keep new surfaces within the ratchet
Follow-through on the post-terminal degrade: the MCP SDK validates
structuredContent against the DECLARED McpRunToolResult schema, so the
new typed `detailProblem` field must exist there (it now does, with the
generated JSON schema regenerated) instead of tripping "additional
properties" and erasing the very result it protects.

Ratchet compliance for the new code, by moving logic to smaller owners
rather than raising any bar: the MCP text/structured result formatters
move from mcp-server/index.ts to run-result-format.ts, and the CLI's
RunDetailProblem descriptor moves from daemon-run.ts (which had crossed
the 600-line cap) into the pure run-detail-projections.ts owner it
belongs with. acp-server keeps its field self-describing.
